Fit CIFs using principal stratum strategy for semicompeting risks data

This function nonparametrically estimates the potential cumulative incidence function using principal stratum strategy (semicompeting risks data structure). The estimand is defined in a subpopulation where intercurrent events would never occur regardless of treatment conditions.

Usage

scr.principal(
  A,
  Time,
  status,
  Time_int,
  status_int,
  weights = rep(1, length(A))
)

Arguments

A

Treatment indicator, 1 for treatment and 0 for control.

Time

Time to the primary (terminal) event.

status

Indicator of the primary (terminal) event, 1 for event and 0 for censoring.

Time_int

Time to the intercurrent event.

status_int

Indicator of the intercurrent event, 1 for event and 0 for censoring.

weights

Weight for each subject.

Value

A list including

time1

Time points in the treated group.

time0

Time points in the control group.

cif1

Estimated cumulative incidence function in the treated group.

cif0

Estimated cumulative incidence function in the control group.

se1

Standard error of the estimated cumulative incidence function in the treated group.

se0

Standard error of the estimated cumulative incidence function in the control group.

time

Time points in both groups.

ate

Estimated treatment effect (difference in cumulative incidence functions).

se

Standard error of the estimated treatment effect.

p.val

P value of testing the treatment effect, which is not available under this strategy.

Details

The principal stratum strategy aims to stratify the population into subpopulations based on the joint potential occurrences of intercurrent events under the two treatment assignments (R(1), R(0)). Suppose we are interested in a principal stratum comprised of individuals who would never experience intercurrent events, regardless of which treatment they receive. This principal stratum can be indicated by \{R(1)=R(0)=\infty\}. The treatment effect is now defined within this subpopulation, \tau(t) = P(T(1) < t \mid R(1)=R(0)=\infty) - P(T(0) < t \mid R(1)=R(0)=\infty), representing the difference in probabilities of experiencing primary outcome events during (0,t) under active treatment and placebo in the subpopulation that will not experience intercurrent events regardless of treatment during (0,t). A principal ignorability assumption is made for identification. If the size of the target principal stratum is small, the results could be highly variable.

See also