Fit CIFs using treatment policy strategy for semicompeting risks data, based on efficient influence functions

This function estimates the potential cumulative incidence function based on efficient influence functions using treatment policy strategy (semicompeting risks data structure). Cox models are employed for the survival model. This strategy ignores the intercurrent event and uses the time to the primary event as it was recorded.

Usage

scr.treatment.eff(A, Time, status, Time_int, status_int, X = NULL)

Arguments

A

Treatment indicator, 1 for treatment and 0 for control.

Time

Time to the primary (terminal) event.

status

Indicator of the primary (terminal) event, 1 for event and 0 for censoring.

Time_int

Time to the intercurrent event.

status_int

Indicator of the intercurrent event, 1 for event and 0 for censoring.

X

Baseline covariates.

Value

A list including

time1

Time points in the treated group.

time0

Time points in the control group.

cif1

Estimated cumulative incidence function in the treated group.

cif0

Estimated cumulative incidence function in the control group.

se1

Standard error of the estimated cumulative incidence function in the treated group.

se0

Standard error of the estimated cumulative incidence function in the control group.

time

Time points in both groups.

ate

Estimated treatment effect (difference in cumulative incidence functions).

se

Standard error of the estimated treatment effect.

p.val

P value of testing the treatment effect based on the efficient influence function of the restricted mean survival time lost by the end of study.

Details

The treatment policy strategy addresses the problem of intercurrent events by expanding the initial treatment conditions to a treatment policy. This strategy is applicable only if intercurrent events do not hinder primary outcome events. The treatments under comparison are now two treatment policies: (w, R(w)), where w = 1, 0. One policy (1,R(1)) involves administering the test drug, along with any naturally occurring intercurrents, whereas the other policy (0,R(0)) involves administering a placebo, along with any naturally occurring intercurrents. Thus, the potential outcomes are T(1,R(1)) and T(0,R(0)). Instead of comparing the test drug and placebo themselves, the contrast of interest is made between the two treatment policies. The difference in cumulative incidences under the two treatment policies is then \tau(t) = P(T(1, R(1)) < t) - P(T(0, R(0)) < t),ATE_tp representing the difference in probabilities of experiencing primary outcome events during (0,t) under active treatment and placebo. The average treatment effect \tau^{\text{tp}}(t) has a meaningful causal interpretation only when T(1, R(1)) and T(0, R(0)) are well defined. Because the treatment policy includes the occurrence of the intercurrent event as natural, the entire treatment policy is determined by manipulating the initial treatment condition $w$ only. Therefore, we can simplify the notations T(w, R(w)) = T(w) in defining estimands. As such, \tau(t) = P(T(1)) < t) - P(T(0) < t) as the intention-to-treat analysis.

See also