Estimate the hazard ratio with intercurrent events

This function estimates the hazard ratio for time-to event data under ICH E9 (R1) to address intercurrent events. Multiple strategies except the principal stratum strategy are allowed.

surv.HR(
  A,
  Time,
  cstatus,
  strategy = "composite",
  cov1 = NULL,
  conf.int = 0.95,
  weights = NULL,
  subset = NULL
)

Arguments

A

Treatment indicator, 1 for treatment and 0 for control.

Time

Time to event.

cstatus

Indicator of event, 1 for the primary event, 2 for the intercurrent event, 0 for censoring.

strategy

Strategy to address intercurrent events, "treatment" indicating treatment policy strategy, "composite" indicating composite variable strategy, "natural" indicating hypothetical strategy (Scenario I, controlling the hazard of intercurrent events), "removed" indicating hypothetical strategy (Scenario II, removing intercurrent events), and "whileon" indicating while on treatment strategy.

cov1

Baseline covariates.

conf.int

Level of the confidence interval.

weights

Weight for each subject (not applied to the while on treatment strategy).

subset

Subset, either numerical or logical.

Value

A list including

logHR

Estimated log hazard ratio (logHR) of the treatment effect on the primary event.

se

Standard error of the estimated log hazard ratio (logHR).

CI

Confidence interval of the hazard ratio (HR).

p.val

P value of the hazard ratio.

Details

For the treatment policy and hypothetical strategies, the hazard ratio (HR) is given by the Cox regression regarding intercurrent events as censoring. For the composite variable strategy, the hazard ratio is given by the Cox regression regarding the first occurrence of either intercurrent event or primary event as the event of interest. For the while on treatment strategy, the hazard ratio is given by the Fine-Gray subdistribution model. There is no existing method to estimate the hazard ratio using principal stratum strategy. The weakness of using hazard ratio to infer treatment effects is critical. First, the hazard ratio relies on model specification. Second, the hazard ratio is not collapsible. Therefore, the hazard ratio should only be treated as a descriptive or exploratory measure of the treatment effect.

Examples

data(bmt)
bmt = transform(bmt, d4=d2+d3)
A = as.numeric(bmt$group>1)
## composite variable strategy
fit = surv.HR(A, bmt$t2, bmt$d4, "composite")
## while on treatment strategy
X = bmt[,c('z1','z3','z5')]
fit = surv.HR(A, bmt$t2, bmt$d4, "whileon", cov1=X)