Fit CIFs using composite variable strategy for competing risks data

This function nonparametrically estimates the potential cumulative incidence function using composite variable strategy (competing risks data structure). This strategy adopts the first occurrence of either the intermediate or primary event as the event of interest.

Usage

surv.composite(A, Time, cstatus, weights = rep(1, length(A)))

Arguments

A

Treatment indicator, 1 for treatment and 0 for control.

Time

Time to event.

cstatus

Indicator of event, 1 for the primary event, 2 for the intercurrent event, 0 for censoring.

weights

Weight for each subject.

Value

A list including

time1

Time points in the treated group.

time0

Time points in the control group.

cif1

Estimated cumulative incidence function in the treated group.

cif0

Estimated cumulative incidence function in the control group.

se1

Standard error of the estimated cumulative incidence function in the treated group.

se0

Standard error of the estimated cumulative incidence function in the control group.

time

Time points in both groups.

ate

Estimated treatment effect (difference in cumulative incidence functions).

se

Standard error of the estimated treatment effect.

p.val

P value of testing the treatment effect based on logrank test.

Details

The composite variable strategy addresses the problem of intercurrent events by expanding the outcome variables. It aggregates the intercurrent event and the primary outcome event into a single composite outcome variable. The idea is not new in the context of progression-free survival, where the composite outcome variable is defined as the occurrence of either a non-terminal event (e.g., cancer progression) or a terminal event (e.g., death). One widely used composite outcome variable has the form Q(w) = \min\{T(w), R(w)\} for w = 1, 0. When this simple form is adopted, the difference in counterfactual cumulative incidences is \tau(t) = P( Q(1) < t ) - P( Q(0) < t ), representing the difference in probabilities of experiencing either intercurrent events or primary outcome events during (0,t) under active treatment and placebo.

See also