Fit CIFs using hypothetical strategy (I) for competing risks data, based on efficient influence functions

This function estimates the potential cumulative incidence function based on efficient influence functions using hypothetical strategy (competing risks data structure). Cox models are employed for survival models. The intercurrent event is only permitted under treated if it would occur under control.

Usage

surv.natural.eff(A, Time, cstatus, X = NULL)

Arguments

A

Treatment indicator, 1 for treatment and 0 for control.

Time

Time to event.

cstatus

Indicator of event, 1 for the primary event, 2 for the intercurrent event, 0 for censoring.

X

Baseline covariates.

Value

A list including

time1

Time points in the treated group.

time0

Time points in the control group.

cif1

Estimated cumulative incidence function in the treated group.

cif0

Estimated cumulative incidence function in the control group.

se1

Standard error of the estimated cumulative incidence function in the treated group.

se0

Standard error of the estimated cumulative incidence function in the control group.

time

Time points in both groups.

ate

Estimated treatment effect (difference in cumulative incidence functions).

se

Standard error of the estimated treatment effect.

p.val

P value of testing the treatment effect based on the efficient influence function of the restricted mean survival time lost by the end of study.

Details

The hypothetical strategy envisions a hypothetical clinical trial condition where the occurrence of intercurrent events is restricted in certain ways. By doing so, the distribution of potential outcomes under the hypothetical scenario can capture the impact of intercurrent events explicitly through a pre-specified criterion. We use T'(w), w = 1, 0 to denote the time to the primary outcome event in the hypothetical scenario. The time-dependent treatment effect specific to this hypothetical scenario is written as \tau(t) = P(T'(1) < t) - P(T'(0) < t), representing the difference in probabilities of experiencing primary outcome events during (0,t) in the pre-specified hypothetical scenario under active treatment and placebo. The key question is how to envision T'(w). We manipulate the hazard specific to intercurrent event \lambda_2(t; w) while assuming the hazard specific to the primary outcome event \lambda_1(t; w) remains unchanged. Specifically, we envision that the intercurrent events that occurred when individuals were assigned to test drugs were only permitted if these intercurrent events would have also occurred if these individuals had been assigned to the placebo. In this hypothetical scenario, when assigned to placebo, individuals would be equally likely to experience intercurrent events as they are assigned to placebo in the real-world trial in terms of the hazards; when assigned to test drug, the hazard of intercurrent events would be identical to that if assigned to placebo in the real-world trial. That is, \lambda_2'(t;0) = \lambda_2'(t;1) = \lambda_2(t;0). The treatment effect corresponds to the natural direct effect with the hazard of intercurrent events set at the level under control.

See also