Baseline hazards of 'tteICE' objects

This function extracts the baseline cumulative hazards in the survival models

Usage

# S3 method for class 'tteICE'
bshaz(x)

Arguments

x

A fitted object returned by the function tteICE, surv.tteICE, or scr.tteICE.

Value

A data frame of baseline cumulative hazards in the working Kaplan-Meier or Cox models, stratified by treatment groups. The first column is time, the following columns are baseline cumulative hazards.

Examples

## load data
data(bmt)
bmt = transform(bmt, d4=d2+d3)
A = as.numeric(bmt$group>1)
X = as.matrix(bmt[,c('z1','z3','z5')])
bmt$A = A

fit = tteICE(Surv(t2, d4, type = "mstate")~A|z1+z3+z5,
 data=bmt, strategy="whileon", method='eff')
bshaz(fit)
#>     time    cumhaz11   cumhaz10    cumhaz21   cumhaz20
#> 1      0 0.000000000 0.00000000 0.000000000 0.00000000
#> 2      1 0.000000000 0.02262243 0.000000000 0.00000000
#> 3      2 0.009558222 0.02262243 0.000000000 0.00000000
#> 4     10 0.019188774 0.02262243 0.000000000 0.00000000
#> 5     16 0.028922275 0.02262243 0.000000000 0.00000000
#> 6     32 0.028922275 0.02262243 0.009956816 0.00000000
#> 7     35 0.038862715 0.02262243 0.009956816 0.00000000
#> 8     47 0.038862715 0.02262243 0.030406963 0.00000000
#> 9     48 0.049059233 0.02262243 0.040761847 0.00000000
#> 10    53 0.059471324 0.02262243 0.040761847 0.00000000
#> 11    55 0.059471324 0.02262243 0.040761847 0.03069172
#> 12    63 0.070016744 0.02262243 0.040761847 0.03069172
#> 13    64 0.070016744 0.02262243 0.051631297 0.03069172
#> 14    74 0.080873693 0.02262243 0.051631297 0.06268677
#> 15    76 0.080873693 0.02262243 0.062806859 0.06268677
#> 16    79 0.092073257 0.02262243 0.062806859 0.06268677
#> 17    80 0.114950586 0.02262243 0.062806859 0.06268677
#> 18    84 0.114950586 0.02262243 0.074421580 0.06268677
#> 19    86 0.114950586 0.04861371 0.074421580 0.06268677
#> 20    93 0.114950586 0.04861371 0.086288562 0.06268677
#> 21   100 0.114950586 0.04861371 0.098347740 0.06268677
#> 22   104 0.114950586 0.04861371 0.098347740 0.09649394
#> 23   105 0.139308434 0.04861371 0.098347740 0.09649394
#> 24   107 0.139308434 0.07575583 0.098347740 0.09649394
#> 25   109 0.139308434 0.07575583 0.098347740 0.13227209
#> 26   110 0.139308434 0.07575583 0.098347740 0.17078441
#> 27   113 0.139308434 0.07575583 0.110817741 0.17078441
#> 28   115 0.139308434 0.07575583 0.123514443 0.17078441
#> 29   120 0.139308434 0.07575583 0.136431405 0.17078441
#> 30   122 0.139308434 0.10443813 0.136431405 0.21252678
#> 31   129 0.139308434 0.10443813 0.136431405 0.26216645
#> 32   157 0.139308434 0.10443813 0.149608189 0.26216645
#> 33   162 0.152471228 0.10443813 0.149608189 0.26216645
#> 34   164 0.165776706 0.10443813 0.149608189 0.26216645
#> 35   168 0.179237790 0.10443813 0.149608189 0.26216645
#> 36   172 0.179237790 0.13572278 0.149608189 0.26216645
#> 37   183 0.192918491 0.13572278 0.149608189 0.26216645
#> 38   192 0.192918491 0.13572278 0.149608189 0.31887495
#> 39   194 0.192918491 0.17129234 0.149608189 0.31887495
#> 40   211 0.192918491 0.17129234 0.163901488 0.31887495
#> 41   219 0.192918491 0.17129234 0.178352463 0.31887495
#> 42   226 0.192918491 0.17129234 0.178352463 0.31887495
#> 43   230 0.192918491 0.17129234 0.178352463 0.38482816
#> 44   242 0.192918491 0.17129234 0.193026578 0.38482816
#> 45   248 0.192918491 0.17129234 0.207851549 0.38482816
#> 46   268 0.192918491 0.17129234 0.222825612 0.38482816
#> 47   272 0.192918491 0.17129234 0.238169638 0.38482816
#> 48   273 0.192918491 0.17129234 0.253862132 0.38482816
#> 49   276 0.192918491 0.21531723 0.253862132 0.38482816
#> 50   288 0.208274796 0.21531723 0.253862132 0.38482816
#> 51   318 0.223953694 0.21531723 0.253862132 0.38482816
#> 52   332 0.223953694 0.26115152 0.253862132 0.38482816
#> 53   363 0.239948323 0.26115152 0.253862132 0.38482816
#> 54   381 0.239948323 0.26115152 0.270594620 0.38482816
#> 55   383 0.239948323 0.26115152 0.270594620 0.46551263
#> 56   390 0.256647984 0.26115152 0.287601419 0.46551263
#> 57   414 0.274000121 0.26115152 0.287601419 0.46551263
#> 58   418 0.274000121 0.31291697 0.287601419 0.46551263
#> 59   421 0.274000121 0.31291697 0.305351438 0.46551263
#> 60   422 0.274000121 0.31291697 0.323332437 0.46551263
#> 61   456 0.274000121 0.31291697 0.341670892 0.46551263
#> 62   466 0.274000121 0.36807667 0.341670892 0.46551263
#> 63   467 0.274000121 0.36807667 0.360358599 0.46551263
#> 64   481 0.292973590 0.36807667 0.360358599 0.46551263
#> 65   486 0.292973590 0.36807667 0.379904130 0.46551263
#> 66   487 0.292973590 0.42644665 0.379904130 0.46551263
#> 67   526 0.292973590 0.49042883 0.379904130 0.46551263
#> 68   530 0.292973590 0.49042883 0.379904130 0.46551263
#> 69   606 0.292973590 0.49042883 0.399896482 0.46551263
#> 70   609 0.292973590 0.49042883 0.399896482 0.55798126
#> 71   625 0.292973590 0.49042883 0.420295677 0.55798126
#> 72   641 0.313670257 0.49042883 0.420295677 0.55798126
#> 73   662 0.313670257 0.49042883 0.420295677 0.66055429
#> 74   677 0.334748991 0.49042883 0.420295677 0.66055429
#> 75   704 0.356179634 0.49042883 0.420295677 0.66055429
#> 76   748 0.356179634 0.49042883 0.442871106 0.66055429
#> 77   845 0.356179634 0.49042883 0.442871106 0.66055429
#> 78   847 0.356179634 0.49042883 0.442871106 0.66055429
#> 79   848 0.356179634 0.49042883 0.442871106 0.66055429
#> 80   860 0.356179634 0.49042883 0.442871106 0.66055429
#> 81   932 0.356179634 0.49042883 0.442871106 0.66055429
#> 82   957 0.356179634 0.49042883 0.442871106 0.66055429
#> 83   996 0.356179634 0.49042883 0.442871106 0.66055429
#> 84  1030 0.356179634 0.49042883 0.442871106 0.66055429
#> 85  1063 0.382486141 0.49042883 0.442871106 0.66055429
#> 86  1074 0.409848291 0.49042883 0.442871106 0.66055429
#> 87  1111 0.409848291 0.49042883 0.442871106 0.66055429
#> 88  1136 0.409848291 0.49042883 0.442871106 0.66055429
#> 89  1167 0.409848291 0.49042883 0.442871106 0.66055429
#> 90  1182 0.409848291 0.49042883 0.442871106 0.66055429
#> 91  1199 0.409848291 0.49042883 0.442871106 0.66055429
#> 92  1238 0.409848291 0.49042883 0.442871106 0.66055429
#> 93  1258 0.409848291 0.49042883 0.442871106 0.66055429
#> 94  1324 0.409848291 0.49042883 0.442871106 0.66055429
#> 95  1330 0.409848291 0.49042883 0.442871106 0.66055429
#> 96  1345 0.409848291 0.49042883 0.442871106 0.66055429
#> 97  1363 0.409848291 0.49042883 0.442871106 0.66055429
#> 98  1377 0.409848291 0.49042883 0.442871106 0.66055429
#> 99  1384 0.409848291 0.49042883 0.442871106 0.66055429
#> 100 1433 0.409848291 0.49042883 0.442871106 0.66055429
#> 101 1447 0.409848291 0.49042883 0.442871106 0.66055429
#> 102 1462 0.409848291 0.49042883 0.442871106 0.66055429
#> 103 1470 0.409848291 0.49042883 0.442871106 0.66055429
#> 104 1496 0.409848291 0.49042883 0.442871106 0.66055429
#> 105 1527 0.409848291 0.49042883 0.442871106 0.66055429
#> 106 1535 0.409848291 0.49042883 0.442871106 0.66055429
#> 107 1562 0.409848291 0.49042883 0.442871106 0.66055429
#> 108 1568 0.409848291 0.49042883 0.442871106 0.66055429
#> 109 1602 0.409848291 0.49042883 0.442871106 0.66055429
#> 110 1631 0.409848291 0.49042883 0.442871106 0.66055429
#> 111 1674 0.409848291 0.49042883 0.442871106 0.66055429
#> 112 1709 0.409848291 0.49042883 0.442871106 0.66055429
#> 113 1799 0.409848291 0.49042883 0.442871106 0.66055429
#> 114 1829 0.409848291 0.49042883 0.442871106 0.66055429
#> 115 1843 0.409848291 0.49042883 0.442871106 0.66055429
#> 116 1850 0.409848291 0.49042883 0.442871106 0.66055429
#> 117 1857 0.409848291 0.49042883 0.442871106 0.66055429
#> 118 1870 0.409848291 0.49042883 0.442871106 0.66055429
#> 119 2024 0.409848291 0.49042883 0.442871106 0.66055429
#> 120 2081 0.409848291 0.49042883 0.442871106 0.66055429
#> 121 2133 0.409848291 0.49042883 0.442871106 0.66055429
#> 122 2140 0.409848291 0.49042883 0.442871106 0.66055429
#> 123 2204 0.523758412 0.49042883 0.442871106 0.66055429
#> 124 2218 0.523758412 0.49042883 0.442871106 0.66055429
#> 125 2246 0.523758412 0.49042883 0.442871106 0.66055429
#> 126 2252 0.523758412 0.49042883 0.442871106 0.66055429
#> 127 2409 0.523758412 0.49042883 0.442871106 0.66055429
#> 128 2430 0.523758412 0.49042883 0.442871106 0.66055429
#> 129 2506 0.523758412 0.49042883 0.442871106 0.66055429
#> 130 2569 0.523758412 0.49042883 0.442871106 0.66055429
#> 131 2640 0.523758412 0.49042883 0.442871106 0.66055429