Print a short summary of the estimated treatment effect

This function summarize the results

# S3 method for class 'tteICE'
print(x, digits = 3, ...)

Arguments

x

A fitted object returned by the function surv.tteICE or scr.tteICE.

digits

The digits of the results

...

Other augments in function print.default

Value

Print the summary of a tteICE object

See also

Examples

## load data
data(bmt)
bmt = transform(bmt, d4=d2+d3)
A = as.numeric(bmt$group>1)
## print the results
for (st in c('composite','natural','removed','whileon','principal')){
 fit = surv.tteICE(A, bmt$t2, bmt$d4, st)
 print(fit)
}
#> The P-value of the estimated treatment effect by strategy composite using np estimation method: 0.591 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>          660   1320   1980   2640
#> CIF1   0.532  0.586  0.586  0.630
#> se1    0.050  0.050  0.050  0.062
#> CIF0   0.609  0.638  0.638  0.638
#> se0    0.080  0.079  0.079  0.079
#> ATE   -0.076 -0.051 -0.051 -0.008
#> se     0.095  0.094  0.094  0.101
#> p.val  0.419  0.584  0.584  0.938
#> The P-value of the estimated treatment effect by strategy natural using np estimation method: 0.638 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>          660   1320   1980  2640
#> CIF1   0.237  0.278  0.278 0.317
#> se1    0.044  0.047  0.047 0.063
#> CIF0   0.311  0.311  0.311 0.311
#> se0    0.078  0.078  0.078 0.078
#> ATE   -0.074 -0.033 -0.033 0.006
#> se     0.086  0.087  0.087 0.095
#> p.val  0.389  0.705  0.705 0.946
#> The P-value of the estimated treatment effect by strategy removed using np estimation method: 0.638 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>          660   1320   1980  2640
#> CIF1   0.280  0.348  0.348 0.417
#> se1    0.050  0.056  0.056 0.084
#> CIF0   0.404  0.404  0.404 0.404
#> se0    0.094  0.094  0.094 0.094
#> ATE   -0.125 -0.056 -0.056 0.012
#> se     0.106  0.109  0.109 0.126
#> p.val  0.239  0.607  0.607 0.921
#> The P-value of the estimated treatment effect by strategy whileon using np estimation method: 0.714 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>          660   1320   1980  2640
#> CIF1   0.239  0.282  0.282 0.324
#> se1    0.043  0.046  0.046 0.062
#> CIF0   0.311  0.311  0.311 0.311
#> se0    0.078  0.078  0.078 0.078
#> ATE   -0.072 -0.028 -0.028 0.013
#> se     0.089  0.090  0.090 0.099
#> p.val  0.419  0.754  0.754 0.898
#> The P-value of the estimated treatment effect by strategy principal using np estimation method: NA 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>          660   1320   1980  2640
#> CIF1   0.344  0.407  0.407 0.466
#> se1    0.058  0.061  0.061 0.083
#> CIF0   0.462  0.462  0.462 0.462
#> se0    0.103  0.103  0.103 0.103
#> ATE   -0.117 -0.055 -0.055 0.005
#> se     0.118  0.120  0.120 0.132
#> p.val  0.320  0.648  0.648 0.972
for (st in c('composite','natural','removed','whileon','principal')){
 fit = scr.tteICE(A, bmt$t1, bmt$d1, bmt$t2, bmt$d2, st)
 print(fit, digits=2)
}
#> The P-value of the estimated treatment effect by strategy composite using np estimation method: 0.59 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>         660  1320  1980  2640
#> CIF1   0.53  0.59  0.59  0.63
#> se1    0.05  0.05  0.05  0.06
#> CIF0   0.61  0.64  0.64  0.64
#> se0    0.08  0.08  0.08  0.08
#> ATE   -0.08 -0.05 -0.05 -0.01
#> se     0.09  0.09  0.09  0.10
#> p.val  0.42  0.58  0.58  0.94
#> The P-value of the estimated treatment effect by strategy natural using np estimation method: 0.5 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>         551  1102  1653  2204
#> CIF1   0.46  0.55  0.57  0.62
#> se1    0.06  0.06  0.06  0.07
#> CIF0   0.55  0.61  0.64  0.64
#> se0    0.08  0.08  0.08  0.08
#> ATE   -0.09 -0.06 -0.07 -0.01
#> se     0.08  0.08  0.08  0.08
#> p.val  0.27  0.46  0.34  0.86
#> The P-value of the estimated treatment effect by strategy removed using np estimation method: 0.64 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>         660  1320  1980 2640
#> CIF1   0.28  0.35  0.35 0.42
#> se1    0.05  0.06  0.06 0.08
#> CIF0   0.40  0.40  0.40 0.40
#> se0    0.09  0.09  0.09 0.09
#> ATE   -0.12 -0.06 -0.06 0.01
#> se     0.11  0.11  0.11 0.13
#> p.val  0.24  0.61  0.61 0.92
#> The P-value of the estimated treatment effect by strategy whileon using np estimation method: 0.71 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>         660  1320  1980 2640
#> CIF1   0.24  0.28  0.28 0.32
#> se1    0.04  0.05  0.05 0.06
#> CIF0   0.31  0.31  0.31 0.31
#> se0    0.08  0.08  0.08 0.08
#> ATE   -0.07 -0.03 -0.03 0.01
#> se     0.09  0.09  0.09 0.10
#> p.val  0.42  0.75  0.75 0.90
#> The P-value of the estimated treatment effect by strategy principal using np estimation method: NA 
#> -------------------------------------------------------------------
#> The estimated cumulative incidences and treatment effects at quartiles:
#>         660  1320  1980 2640
#> CIF1   0.34  0.41  0.41 0.47
#> se1    0.06  0.06  0.06 0.08
#> CIF0   0.46  0.46  0.46 0.46
#> se0    0.10  0.10  0.10 0.10
#> ATE   -0.12 -0.05 -0.05 0.00
#> se     0.12  0.12  0.12 0.13
#> p.val  0.32  0.65  0.65 0.97