tteICE: Treatment Effect Estimation for Time-to-Event Data with Intercurrent Events

Main functions:

tteICE Fit the CIF for time-to-event data using formula

scr.tteICE Fit the CIF for time-to-event data with intercurrent events for semicompeting risks data

surv.tteICE Fit the CIF for time-to-event with intercurrent events for competing risks data

plot.tteICE Graphical results of tteICE

print.tteICE Print a short summary of the estimated treatment effect

summary.tteICE Summary of tteICE

predict.tteICE Risk prediction at specific time points

tteICEShiny Interactive Shiny app for tteICE

Example data:

bmt Data from Section 1.3 of Klein and Moeschberger (1997)

See also

Author

Maintainer: Yi Zhou yzhou@pku.edu.cn

Authors:

Yuhao Deng dengyuhao@pku.edu.cn